+++ Ship.py
@@ -1,0 +1,8 @@
+
+class Ship(object):
+    def __init__(self, x, y):
+        self.x = x
+        self.y = y
+
+    def __repr__(self):
+        return "(%d, %d)" % (self.x, self.y)
